% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{format_t}
\alias{format_t}
\title{Format T-Test}
\usage{
format_t(fit, rounding = 3)
}
\arguments{
\item{fit}{a t-test object}

\item{rounding}{the number of decimal places for the output}
}
\value{
a data.frame with useful t-test information
}
\description{
Formats a t-test object to look gorgeous.
}
